Create a common progression component
