+++ components/map-sidebar/index.js
@@ -1,14 +1,8 @@
 import PropTypes from 'prop-types'
 import {useRouter} from 'next/router'
-import {find} from 'lodash'
 
-import {formatDate} from '@/lib/date-utils.js'
-import {findClosestEtape} from '@/shared/find-closest-etape.js'
-import {STATUS} from '@/lib/utils/projet.js'
-
+import Progression from '@/components/map-sidebar/progression.js'
 import Header from '@/components/map-sidebar/project-header.js'
-import Badge from '@/components/badge.js'
-import Timeline from '@/components/map-sidebar/timeline.js'
 import PcrsInfos from '@/components/map-sidebar/pcrs-infos.js'
 import Documents from '@/components/map-sidebar/documents.js'
 import Contact from '@/components/map-sidebar/contact.js'
@@ -36,12 +30,6 @@ const MapSidebar = ({projet, onClose, onProjetChange, projets}) => {
   } = projet
 
   const contactAPLC = acteurs.find(acteur => acteur.role === 'aplc' || 'porteur')
-  const {statut} = projet.etapes[projet.etapes.length - 1]
-  const projectStartDate = formatDate(find(projet.etapes, {statut: 'investigation'}).date_debut)
-  const isObsolete = statut === 'obsolete'
-
-  const closestPostStep = findClosestEtape(etapes)
-  const closestPostStepStatus = STATUS[closestPostStep.statut]
 
   return (
     <>
@@ -55,25 +43,7 @@ const MapSidebar = ({projet, onClose, onProjetChange, projets}) => {
       />
       <div className='infos-container'>
         <h2 className='fr-text--lead fr-mb-1w'>État d’avancement</h2>
-        <div className='actual-status fr-mb-3w'>
-          <Badge
-            background={closestPostStepStatus.color}
-            textColor={closestPostStepStatus.textColor}
-          >
-            {closestPostStepStatus.label}
-          </Badge>
-
-          {projectStartDate && (
-            <div className='start-date fr-text--sm fr-m-0'>Lancement du projet le {projectStartDate}</div>
-          )}
-        </div>
-        {!isObsolete && (
-          <Timeline
-            stepsColors={status}
-            currentStatus={closestPostStep.statut}
-            steps={etapes}
-          />
-        )}
+        <Progression etapes={etapes} />
 
         <h2 className='fr-text--lead fr-my-0'>Détails du projet</h2>
         <PcrsInfos
@@ -115,14 +85,6 @@ const MapSidebar = ({projet, onClose, onProjetChange, projets}) => {
         .infos-container {
           padding: 1em;
           background-color: #ffffff;
-        }
-        .actual-status {
-          display: flex;
-          flex-wrap: wrap;
-          gap: 10px;
-        }
-        .start-date {
-          font-style: italic;
         }
       `}</style>
     </>
